refactor(factory): separate DummyPlayerFactory to test package
